feat: getTransactionInfo, saveAllowanceRequest, getScheduledAllowance 구현

+++ back-end/chatbot/src/main/java/com/aicheck/chatbot/domain/chat/TransactionRecord.java
@@ -5,6 +5,6 @@ public record TransactionRecord(
 	String secondCategoryName,
 	TransactionType transactionType,
 	Integer amount,
-	Short rating
+	Integer rating
 ) {
 }
